Prohibit implicit narrowing type conversions
In some parts it might be overdone, but helped to find a few bugs/quirks
which otherwise probably would have not been found in time.

+++ src/providers/cma/connection.hpp
@@ -38,7 +38,7 @@ namespace pMR
                 Connection& operator=(Connection&&) = delete;
                 ~Connection() = default;
                 void sendAddress(iovec &buffer) const;
-                void sendData(iovec buffer, std::uint32_t const sizeByte) const;
+                void sendData(iovec buffer, std::size_t const sizeByte) const;
                 void writeData(iovec localBuffer, iovec remoteBuffer) const;
                 void postNotifySend() const;
                 void pollNotifySend();
